refactor: improve scope and add to task node
