+++ manifest_meta/src/node.rs
@@ -2,7 +2,10 @@ use std::{collections::HashMap, path::PathBuf, sync::Arc};
 
 use manifest_reader::manifest::{InputDefPatch, InputHandles, OutputHandles};
 
-use crate::{Block, HandleName, NodeId, ServiceBlock, SlotBlock, SubflowBlock, TaskBlock};
+use crate::{
+    scope::RunningScope, Block, HandleName, NodeId, ServiceBlock, SlotBlock, SubflowBlock,
+    TaskBlock,
+};
 
 pub type HandlesFroms = HashMap<HandleName, Vec<HandleFrom>>;
 
@@ -31,6 +34,7 @@ macro_rules! extend_node_common_field {
 extend_node_common_field!(TaskNode {
     task: Arc<TaskBlock>,
     timeout_seconds: Option<u64>,
+    scope: RunningScope,
 });
 
 extend_node_common_field!(ServiceNode {
